requirements.txt
streamlit>=1.28.0
langchain>=0.2.0
langchain-openai>=0.1.20
langchain-community>=0.2.0
langchain-core>=0.2.40
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0
openai>=1.40.0
python-dotenv>=1.0.0
lxml>=4.9.0
html2text>=2020.1.16
urllib3>=2.1.0
pandas>=2.1.0
numpy>=1.24.0
tiktoken>=0.7.0
